UI Particles
All working as intended, potential tweaks going forward to the appearance of each.
NOTE: The particle UI does not disable when paused, needs the UI scripting similar to other UI canvas.
All UI particles are ready to be integrated with the game states and mechanics as appropraite to each particle system.